refactor: moved scaling and resizing to separate cog
