+++ main.py
@@ -7,7 +7,7 @@ from discord.ext import commands
 from image_utils import process_command, spoof_human
 from cogs.color_commands import ColorCog
 from utils import color
-from utils import scaling
+from utils import transformation
 
 intents = discord.Intents.default()
 intents.message_content = True
@@ -39,57 +39,6 @@ async def image_error_handler(ctx, error):
     else:
         await ctx.send(f"Something unexpected happened: {error}")
 
-@bot.command(name="scale", description="scale image by factor")
-async def scale_image(ctx, factor, *args):
-    async def scaling_wrapper(img_path, factor):
-        display = scaling.image_scaling(img_path, factor)
-        if not display:
-            await ctx.send("To see the image, please copy the link and open it in a browser")
-    await process_command(ctx, scaling_wrapper, *args, factor=factor)
-
-@scale_image.error
-async def scale_error_handler(ctx, error):
-    if isinstance(error, commands.MissingRequiredArgument) or isinstance(error, commands.TooManyArguments):
-        await ctx.send("Usage: $scale [factor] [url]")
-    elif isinstance(error, commands.BadArgument):
-        await ctx.send("Factor needs to be a real positive number")
-    elif isinstance(error, commands.UserInputError):
-        await ctx.send("Factor is either too small or too big. Please choose an appropriate factor")
-    else:
-        await ctx.send(f"Something unexpected happened: {error}")
-
-@bot.command(name="resize", description="resize image by width and height")
-async def resize_image(ctx, width, height, *args):
-    async def resize_image_wrapper(img_path, width, height):
-        display = scaling.image_resizing(img_path, width, height)
-        if not display:
-            await ctx.send("To see the image, please copy the link and open it in a browser")
-    await process_command(ctx, resize_image_wrapper, *args, width=width, height=height)
-
-@resize_image.error
-async def resize_error_handler(ctx, error):
-    if isinstance(error, commands.MissingRequiredArgument) or isinstance(error, commands.TooManyArguments):
-        await ctx.send("Usage: $resize [width] [height] [url]")
-    elif isinstance(error, commands.BadArgument):
-        await ctx.send("Width and Height need to be positive integers less than or equal to 65500")
-    elif isinstance(error, commands.UserInputError):
-        await ctx.send("Resized image exceeds file size limit. Please choose smaller width and height")
-    else:
-        await ctx.send(f"Something unexpected happened: {error}")
-
-"""
-@bot.command(name="grayscale", description="Test that the bot can download images and send them back converted to grayscale")
-async def grayscale(ctx, *args):
-    await process_command(ctx, color.grayscale, *args)
-
-@grayscale.error
-async def grayscale_error_handler(ctx, error):
-    if isinstance(error, commands.MissingRequiredArgument):
-        await ctx.send("Please send a URL linking to your image")
-    else:
-        await ctx.send(f"Something unexpected happened: {error}")
-"""
-
 if __name__ == "__main__":
     # heroku - get token from environment variable
     if "IMAGE_BOT_TOKEN" in os.environ:
@@ -102,5 +51,8 @@ if __name__ == "__main__":
         raise RuntimeError("Could not find token")
 
     spoof_human()
-    asyncio.run(bot.load_extension("cogs.color_commands"))
+
+    cogs = ["color", "transformation"]
+    for cog in cogs:
+        asyncio.run(bot.load_extension(f"cogs.{cog}_commands"))
     bot.run(BOT_TOKEN)
